# Optimize __getitem__ for boolean array indices