@rezoo 2017-01-26
Thank you for your PR!
As far as I can see, removing the `cupy_max` kernel from `__getitem__` has the biggest effect on the reduction of the computational time.