@rezoo 2017-01-26
Good. I think all the points you have modified seem to be valid. LGTM.